src/player: Rewrite internal time player standard
* src/formatter/FormatterMediator.cpp
* src/formatter/FormatterMediator.h
* src/formatter/NominalEventMonitor.cpp
* src/player/AVPlayer.cpp
* src/player/AVPlayer.h
* src/player/ChannelPlayer.cpp
* src/player/ChannelPlayer.h
* src/player/INCLPlayer.h
* src/player/IPlayer.h
* src/player/Player.cpp
* src/player/Player.h

+++ src/formatter/FormatterMediator.h
@@ -321,9 +321,9 @@ public:
   void setSurface (SDLSurface* surface);
   SDLSurface* getSurface ();
   void flip ();
-  double getMediaTime ();
+  guint32 getMediaTime ();
   double getTotalMediaTime ();
-  void setMediaTime (double newTime);
+  void setMediaTime (guint32 newTime);
   bool setKeyHandler (bool isHandler);
   void setScope (const string &scope, short type, double begin = -1,
                  double end = -1, double outTransDur = -1);
